Refactor tree placement filters into PositionFilters
